refactor(mf): clock - scss into css
