+++ micro-frontends/src/mf-clock/src/ui/modules/time/time.module.css
@@ -1,0 +1,7 @@
+.time {
+	color: var(--ds-text-color-primary);
+	font-family: var(--ds-font-sans-serif-medium);
+	font-size: var(--ds-font-size-xs);
+	line-height: var(--ds-line-height-normal);
+	letter-spacing: 0.7px;
+}
