+++ micro-frontends/src/mf-clock/src/ui/modules/time/time.tsx
@@ -1,6 +1,6 @@
 import { observer } from "mobx-react-lite"
 import type { FunctionComponent } from "react"
-import styles from "./time.module.scss"
+import styles from "./time.module.css"
 import { useTime } from "./use-time.hook"
 
 export const Time: FunctionComponent = observer(() => {
