+++ micro-frontends/src/mf-clock/src/ui/modules/wrapper/wrapper.module.css
@@ -1,0 +1,4 @@
+.wrapper {
+	display: flex;
+	background-color: var(--ds-background-color-primary);
+}
